+++ common/src/main/java/com/slymask3/instantblocks/block/instant/InstantLavaBlock.java
@@ -16,7 +16,7 @@ import net.minecraft.world.level.material.Material;
 
 public class InstantLavaBlock extends InstantLiquidBlock {
 	public InstantLavaBlock() {
-		super(Block.Properties.of(Material.LAVA)
+		super(Block.Properties.of(Material.CLAY)
 				.strength(0.5F)
 				.sound(new LiquidSoundType(SoundEvents.BUCKET_FILL_LAVA, SoundEvents.BUCKET_EMPTY_LAVA))
 				.lightLevel((par1) -> 15)